Add CompoundOreDict Move IOreDict
