+++ src/main/java/io/ib67/astralflow/manager/ItemRegistry.java
@@ -21,8 +21,12 @@
 
 package io.ib67.astralflow.manager;
 
-import io.ib67.astralflow.item.*;
+import io.ib67.astralflow.item.AstralItem;
+import io.ib67.astralflow.item.ItemKey;
+import io.ib67.astralflow.item.ItemState;
+import io.ib67.astralflow.item.StateScope;
 import io.ib67.astralflow.item.factory.ItemPrototypeFactory;
+import io.ib67.astralflow.item.oredict.IOreDict;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;
