+++ src/main/java/io/ib67/astralflow/manager/impl/ItemRegistryImpl.java
@@ -24,8 +24,12 @@ package io.ib67.astralflow.manager.impl;
 import io.ib67.astralflow.AstralFlow;
 import io.ib67.astralflow.hook.HookType;
 import io.ib67.astralflow.internal.item.state.InternalItemState;
-import io.ib67.astralflow.item.*;
+import io.ib67.astralflow.item.AstralItem;
+import io.ib67.astralflow.item.ItemKey;
+import io.ib67.astralflow.item.ItemState;
+import io.ib67.astralflow.item.StateScope;
 import io.ib67.astralflow.item.factory.ItemPrototypeFactory;
+import io.ib67.astralflow.item.oredict.IOreDict;
 import io.ib67.astralflow.item.tag.UUIDTag;
 import io.ib67.astralflow.manager.ItemRegistry;
 import io.ib67.astralflow.storage.ItemStateStorage;
